standardize export formatting
