+++ packages/react-components/theme-designer/src/utils/toString.ts
@@ -1,0 +1,41 @@
+import type { BrandVariants, Theme } from '@fluentui/react-components';
+import { brandRamp } from '../components/ColorTokens/getOverridableTokenBrandColors';
+
+export const getBrandValues = (brand: BrandVariants, overrideList: Partial<Theme>, spacer: string) => {
+  const hexToBrand: Record<string, string> = {};
+
+  brandRamp.map(i => {
+    hexToBrand[brand[i]] = 'brand[' + `${i as number}` + ']';
+    return;
+  });
+
+  const stringOverrideList = overrideList as Record<string, string>;
+
+  return (
+    Object.keys(overrideList).map(token => {
+      return '\n' + spacer + token + ': ' + hexToBrand[stringOverrideList[token]];
+    }) + '\n'
+  );
+};
+
+export const objectToString = (input: Record<string, string>, spacer: string) => {
+  return (
+    Object.keys(input).map(key => {
+      return '\n' + spacer + key + ": '" + input[key] + "'";
+    }) + '\n'
+  );
+};
+
+export const themeToString = (theme: Theme, spacer: string) => {
+  const indexableTheme: Record<string, any> = theme;
+  return (
+    // don't put quotes around numbers
+    Object.keys(theme).map(key => {
+      if (!Number(indexableTheme[key])) {
+        return '\n' + spacer + key + ": '" + indexableTheme[key] + "'";
+      } else {
+        return '\n' + spacer + key + ': ' + indexableTheme[key];
+      }
+    }) + '\n'
+  );
+};
